[Melhoria] Troca do filterBy por groupBy
